WS-002: Change Int to Float on Lat and Long and remove blank lines

+++ backend/Models/Database/Location.cs
@@ -2,18 +2,13 @@ using System.Collections.Generic;
 
 namespace WhaleSpotting.Models.Database
 {
-
     public class Location
     {
         public int Id { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public List<string> Amenities { get; set; }
-
-        
-
     }
-
 }
